+++ listeners/messages/index.js
@@ -1,3 +1,0 @@
-export const register = (_app) => {
-  // No message listeners needed for the flag translator
-};